requirements.txt
fastapi
uvicorn
torch
ftfy
regex
tqdm
Pillow
git+https://github.com/openai/CLIP.git
